Fix Windows Build Warnings
This PR resolves Windows build warnings.

- Type Conversion Warnings (C4244, C4267, C4146, C4293)
- MSVC Compiler Flag Warnings (D9024, D9025, D9027)
- Hexadecimal Floating-Point Warnings (C4305, C4309)

**Self-evaluation:**
1. Build test: [X]Passed [ ]Failed [ ]Skipped
2. Run test:   [X]Passed [ ]Failed [ ]Skipped

+++ nntrainer/layers/embedding.cpp
@@ -187,7 +187,7 @@ void EmbeddingLayer::calcGradient(RunLayerContext &context) {
 
     if (djdw.getDataType() == TensorDim::DataType::FP32) {
       for (unsigned int i = 0; i < input_.width(); ++i) {
-        unsigned int embed_idx = ((float *)(in_data))[i];
+        unsigned int embed_idx = (unsigned int)((float *)(in_data))[i];
         // Assume padding is 0 and index always start from 1.
         // If in_data[i] - 1 < 0, then it skips.
         // if (embed_idx == 0)
@@ -203,7 +203,7 @@ void EmbeddingLayer::calcGradient(RunLayerContext &context) {
     } else if (djdw.getDataType() == TensorDim::DataType::FP16) {
 #ifdef ENABLE_FP16
       for (unsigned int i = 0; i < input_.width(); ++i) {
-        unsigned int embed_idx = ((float *)(in_data))[i];
+        unsigned int embed_idx = (unsigned int)((float *)(in_data))[i];
         // Assume padding is 0 and index always start from 1.
         // If in_data[i] - 1 < 0, then it skips.
         // if (embed_idx == 0)
